requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
faiss-cpu
langchain_google_genai
langchain-astradb

ollama

pdfplumber 
langchain-core
langchain-community
langchain_text_splitters

groq
langchain-groq
google-cloud-aiplatform>=1.38

sentence-transformers
